backend/requirements.txt
# Backend requirements for Campus Hub
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
PyMySQL==1.1.0
PyJWT==2.10.1

# Additional for production
gunicorn==21.2.0
